Add files via upload
These are temporary files! Some more edits and organization needs to be made to them. It should be good to test the code.

Cuisine Map - draw all features, but then for each cuisine feature - no fill and no stroke. My view is that we have the globe with all the countries with the regions overlaid as necessary in a different color and opacity of 0.4. 

For legend SVG, I am thinking a pull down menu to select the spice of interest, and then the cuisines become bolded. Does this sound good?